% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-package.R
\docType{package}
\name{cps-package}
\alias{cps}
\alias{cps-package}
\title{cps: Reproducible Continuous PD-L1 CPS Modeling (AGAMENON)}
\description{
Data and helper functions to reproduce the tables and figures of the
AGAMENON analysis of PD-L1 CPS and first-line immunotherapy in advanced
gastric/gastroesophageal junction adenocarcinoma.
}
\details{
Main functions: \code{\link{cps_prepare}}, \code{\link{cps_table1}},
\code{\link{cps_distribution}}, \code{\link{cps_fit}},
\code{\link{cps_table2}}, \code{\link{cps_model_comparison}},
\code{\link{cps_figure_hr}}.
}
\author{
\strong{Maintainer}: Alberto Carmona-Bayonas \email{alberto.carmonabayonas@gmail.com} (Medical Oncology Department, Hospital Universitario Morales Meseguer, Murcia, Spain)

Authors:
\itemize{
  \item AGAMENON-SEOM Investigators
}

}
\keyword{internal}
