% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cps_clean}
\alias{cps_clean}
\title{Clean IPUMS CPS data}
\usage{
cps_clean(data)
}
\arguments{
\item{data}{Data frame of IPUMS CPS data.}
}
\value{
Data frame.
}
\description{
Currently, this only cleans/recodes a selection of variables.
}
