% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cps_db_list}
\alias{cps_db_list}
\title{List variables in IPUMS CPS database table}
\usage{
cps_db_list(file_db)
}
\arguments{
\item{file_db}{Path of database file to read.}
}
\value{
Character vector of variable names from database table.
}
\description{
List variables in IPUMS CPS database table
}
