% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cps_db_read}
\alias{cps_db_read}
\title{Read IPUMS CPS database table.}
\usage{
cps_db_read(file_db, years, vars)
}
\arguments{
\item{file_db}{Path of database file to read.}

\item{years}{Years in data to query (where).}

\item{vars}{Variables in data to query (select).}
}
\value{
Data frame.
}
\description{
Read IPUMS CPS database table.
}
