% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{cps_db_write}
\alias{cps_db_write}
\title{Write IPUMS CPS database table}
\usage{
cps_db_write(file_data, file_db)
}
\arguments{
\item{file_data}{Path of CSV file of IPUMS CPS data.}

\item{file_db}{Path of database to create.}
}
\description{
Write database table from IPUMS CPS CSV file. This will overwrite the
existing database file with the same name.
}
