% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{cps_model_comparison}
\alias{cps_model_comparison}
\title{Supplementary Table S4 - model comparison (AIC, C-index, R2)}
\usage{
cps_model_comparison(object = cps_fit())
}
\arguments{
\item{object}{A fitted object from \code{\link{cps_fit}}.}
}
\value{
A data frame comparing the spline, interval and binary models.
}
\description{
Supplementary Table S4 - model comparison (AIC, C-index, R2)
}
