% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{cps_table1}
\alias{cps_table1}
\title{Table 1 - baseline characteristics by treatment arm}
\usage{
cps_table1(data = cps::agamenon_cps)
}
\arguments{
\item{data}{A data frame like \code{agamenon_cps}.}
}
\value{
A \code{gtsummary} table.
}
\description{
Table 1 - baseline characteristics by treatment arm
}
