% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-package.R
\docType{package}
\name{cps-package}
\alias{cps}
\alias{cps-package}
\title{cps: Tools for Working with IPUMS CPS Data}
\description{
Tools for working with IPUMS CPS data.
}
\author{
\strong{Maintainer}: Ray Wright \email{raywright88@gmail.com}

}
\keyword{internal}
